hazelnut-insurance/requirements.txt
cdsapi>=0.7.7
xarray>=2024.0.0
netCDF4>=1.6.0
h5netcdf>=1.3.0
faostat>=2.1.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
scipy>=1.11.0
jupyter>=1.0.0
notebook>=7.0.0
pytest>=7.0.0
requests>=2.31.0
tqdm>=4.66.0
